tests/wpt/web-platform-tests/css/css-box/parsing/margin-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS basic box model: getComputedStyle().margin</title>
<link rel="help" href="https://drafts.csswg.org/css-box-3/#propdef-margin">
<meta name="assert" content="margin computed value has absolute lengths.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
  #parent {
    width: 200px;
  }
  #target {
    width: 0px;
    font-size: 40px;
  }
</style>
</head>
<body>
<div id="parent">
  <div id="target"></div>
</div>
<script>
test_computed_value("margin", "10px");
test_computed_value("margin", "10px 20px 30px 40px");
test_computed_value("margin", "calc(0.5em + 10px)", "30px");
test_computed_value("margin", "30%", "60px");

test_computed_value("margin-top", "10px");
test_computed_value("margin-right", "20px");
test_computed_value("margin-bottom", "30px");
test_computed_value("margin-left", "40px");
</script>
</body>
</html>
